youtube_insights/requirements.txt
yt-dlp
openai-whisper
torch
torchvision
torchaudio
opencv-python
transformers
Pillow
numpy
python-dotenv
